Separate json and yaml loaders
