--- test/test_goodway_configs/data/config.toml
@@ -1,2 +1,0 @@
-key1 = "value1"
-key2 = 12345
